Activity checkboxes
Did a quick fix on the payment methods area, as I was hiding it on load when I shouldn't have been when I looked over the insctructions again.
activities will also now add to a total cost when checked (and removes that additional cost when unchecked) as well as disabling the checkboxes that have a conflicting time with one another when one of them is checked.